+++ tests/streamlit/test_commands.py
@@ -55,9 +55,9 @@ def test_describe_streamlit(mock_connector, runner, mock_ctx):
     ]
 
 
-def _put_query(source: str, dest: str):
+def _put_query(project_root: Path, source: str, dest: str):
     return dedent(
-        f"put file://{Path(source)} {dest} auto_compress=false parallel=4 overwrite=True"
+        f"put file://{project_root.resolve() / 'output' / source} {dest} auto_compress=false parallel=4 overwrite=True"
     )
 
 
@@ -91,16 +91,18 @@ def test_deploy_only_streamlit_file(
     mock_connector.return_value = ctx
     mock_get_account.return_value = "my_account"
 
-    with project_directory("example_streamlit") as pdir:
-        (pdir / "environment.yml").unlink()
-        shutil.rmtree(pdir / "pages")
+    with project_directory("example_streamlit") as tmp_dir:
+        (tmp_dir / "environment.yml").unlink()
+        shutil.rmtree(tmp_dir / "pages")
         result = runner.invoke(["streamlit", "deploy"])
 
     assert result.exit_code == 0, result.output
     assert ctx.get_queries() == [
         "create stage if not exists IDENTIFIER('MockDatabase.MockSchema.streamlit')",
         _put_query(
-            "streamlit_app.py", "@MockDatabase.MockSchema.streamlit/test_streamlit"
+            tmp_dir,
+            "streamlit_app.py",
+            "@MockDatabase.MockSchema.streamlit/test_streamlit",
         ),
         dedent(
             f"""
@@ -146,16 +148,18 @@ def test_deploy_only_streamlit_file_no_stage(
     mock_connector.return_value = ctx
     mock_get_account.return_value = "my_account"
 
-    with project_directory("example_streamlit_no_stage") as pdir:
-        (pdir / "environment.yml").unlink()
-        shutil.rmtree(pdir / "pages")
+    with project_directory("example_streamlit_no_stage") as tmp_dir:
+        (tmp_dir / "environment.yml").unlink()
+        shutil.rmtree(tmp_dir / "pages")
         result = runner.invoke(["streamlit", "deploy"])
 
     assert result.exit_code == 0, result.output
     assert ctx.get_queries() == [
         "create stage if not exists IDENTIFIER('MockDatabase.MockSchema.streamlit')",
         _put_query(
-            "streamlit_app.py", "@MockDatabase.MockSchema.streamlit/test_streamlit"
+            tmp_dir,
+            "streamlit_app.py",
+            "@MockDatabase.MockSchema.streamlit/test_streamlit",
         ),
         dedent(
             f"""
@@ -200,18 +204,22 @@ def test_deploy_with_empty_pages(
     mock_connector.return_value = ctx
     mock_get_account.return_value = "my_account"
 
-    with project_directory("streamlit_empty_pages") as directory:
-        (directory / "pages").mkdir(parents=True, exist_ok=True)
+    with project_directory("streamlit_empty_pages") as tmp_dir:
+        (tmp_dir / "pages").mkdir(parents=True, exist_ok=True)
         result = runner.invoke(["streamlit", "deploy"])
 
     assert result.exit_code == 0, result.output
     assert ctx.get_queries() == [
         "create stage if not exists IDENTIFIER('MockDatabase.MockSchema.streamlit')",
         _put_query(
-            "streamlit_app.py", "@MockDatabase.MockSchema.streamlit/test_streamlit"
+            tmp_dir,
+            "streamlit_app.py",
+            "@MockDatabase.MockSchema.streamlit/test_streamlit",
         ),
         _put_query(
-            "environment.yml", "@MockDatabase.MockSchema.streamlit/test_streamlit"
+            tmp_dir,
+            "environment.yml",
+            "@MockDatabase.MockSchema.streamlit/test_streamlit",
         ),
         dedent(
             f"""
@@ -223,7 +231,6 @@ def test_deploy_with_empty_pages(
         ).strip(),
         "select system$get_snowsight_host()",
     ]
-    assert "Skipping empty directory: pages" in result.output
 
 
 @mock.patch("snowflake.cli._plugins.connection.util.get_account")
@@ -256,16 +263,18 @@ def test_deploy_only_streamlit_file_replace(
     mock_connector.return_value = ctx
     mock_get_account.return_value = "my_account"
 
-    with project_directory("example_streamlit") as pdir:
-        (pdir / "environment.yml").unlink()
-        shutil.rmtree(pdir / "pages")
+    with project_directory("example_streamlit") as tmp_dir:
+        (tmp_dir / "environment.yml").unlink()
+        shutil.rmtree(tmp_dir / "pages")
         result = runner.invoke(["streamlit", "deploy", "--replace"])
 
     assert result.exit_code == 0, result.output
     assert ctx.get_queries() == [
         "create stage if not exists IDENTIFIER('MockDatabase.MockSchema.streamlit')",
         _put_query(
-            "streamlit_app.py", "@MockDatabase.MockSchema.streamlit/test_streamlit"
+            tmp_dir,
+            "streamlit_app.py",
+            "@MockDatabase.MockSchema.streamlit/test_streamlit",
         ),
         dedent(
             f"""
@@ -279,23 +288,6 @@ def test_deploy_only_streamlit_file_replace(
         "select system$get_snowsight_host()",
     ]
     mock_typer.launch.assert_not_called()
-
-
-def test_artifacts_must_exist(
-    runner, mock_ctx, project_directory, alter_snowflake_yml, snapshot
-):
-    with project_directory("example_streamlit_v2") as pdir:
-        alter_snowflake_yml(
-            pdir / "snowflake.yml",
-            parameter_path="entities.my_streamlit.artifacts.1",
-            value="foo_bar.py",
-        )
-
-        result = runner.invoke(
-            ["streamlit", "deploy"],
-        )
-        assert result.exit_code == 1
-        assert result.output == snapshot
 
 
 @pytest.mark.parametrize("project_name", ["example_streamlit_v2", "example_streamlit"])
@@ -364,11 +356,11 @@ def test_deploy_streamlit_and_environment_files(
     )
     mock_connector.return_value = ctx
 
-    with project_directory(project_name) as pdir:
-        shutil.rmtree(pdir / "pages")
+    with project_directory(project_name) as tmp_dir:
+        shutil.rmtree(tmp_dir / "pages")
         if project_name == "example_streamlit_v2":
             alter_snowflake_yml(
-                pdir / "snowflake.yml",
+                tmp_dir / "snowflake.yml",
                 parameter_path="entities.test_streamlit.artifacts",
                 value=["streamlit_app.py", "environment.yml"],
             )
@@ -379,8 +371,8 @@ def test_deploy_streamlit_and_environment_files(
     assert result.exit_code == 0, result.output
     assert ctx.get_queries() == [
         "create stage if not exists IDENTIFIER('MockDatabase.MockSchema.streamlit')",
-        _put_query("streamlit_app.py", root_path),
-        _put_query("environment.yml", root_path),
+        _put_query(tmp_dir, "streamlit_app.py", root_path),
+        _put_query(tmp_dir, "environment.yml", root_path),
         dedent(
             f"""
             CREATE STREAMLIT IDENTIFIER('MockDatabase.MockSchema.{STREAMLIT_NAME}')
@@ -423,11 +415,11 @@ def test_deploy_streamlit_and_pages_files(
     )
     mock_connector.return_value = ctx
 
-    with project_directory(project_name) as pdir:
-        (pdir / "environment.yml").unlink()
+    with project_directory(project_name) as tmp_dir:
+        (tmp_dir / "environment.yml").unlink()
         if project_name == "example_streamlit_v2":
             alter_snowflake_yml(
-                pdir / "snowflake.yml",
+                tmp_dir / "snowflake.yml",
                 parameter_path="entities.test_streamlit.artifacts",
                 value=["streamlit_app.py", "pages/"],
             )
@@ -437,8 +429,8 @@ def test_deploy_streamlit_and_pages_files(
     assert result.exit_code == 0, result.output
     assert ctx.get_queries() == [
         "create stage if not exists IDENTIFIER('MockDatabase.MockSchema.streamlit')",
-        _put_query("streamlit_app.py", root_path),
-        _put_query("pages/*", f"{root_path}/pages"),
+        _put_query(tmp_dir, "streamlit_app.py", root_path),
+        _put_query(tmp_dir, "pages/my_page.py", f"{root_path}/pages"),
         dedent(
             f"""
             CREATE STREAMLIT IDENTIFIER('MockDatabase.MockSchema.{STREAMLIT_NAME}')
@@ -482,18 +474,18 @@ def test_deploy_all_streamlit_files(
     )
     mock_connector.return_value = ctx
 
-    with project_directory(project_name):
+    with project_directory(project_name) as tmp_dir:
         result = runner.invoke(["streamlit", "deploy"])
 
     root_path = f"@MockDatabase.MockSchema.streamlit/{STREAMLIT_NAME}"
     assert result.exit_code == 0, result.output
     assert ctx.get_queries() == [
         "create stage if not exists IDENTIFIER('MockDatabase.MockSchema.streamlit')",
-        _put_query("streamlit_app.py", root_path),
-        _put_query("environment.yml", root_path),
-        _put_query("pages/*", f"{root_path}/pages"),
-        _put_query("utils/utils.py", f"{root_path}/utils"),
-        _put_query("extra_file.py", root_path),
+        _put_query(tmp_dir, "streamlit_app.py", root_path),
+        _put_query(tmp_dir, "environment.yml", root_path),
+        _put_query(tmp_dir, "pages/my_page.py", f"{root_path}/pages"),
+        _put_query(tmp_dir, "utils/utils.py", f"{root_path}/utils"),
+        _put_query(tmp_dir, "extra_file.py", root_path),
         dedent(
             f"""
             CREATE STREAMLIT IDENTIFIER('MockDatabase.MockSchema.{STREAMLIT_NAME}')
@@ -555,16 +547,16 @@ def test_deploy_put_files_on_stage(
     with project_directory(
         project_name,
         merge_project_definition=merge_definition,
-    ):
+    ) as tmp_dir:
         result = runner.invoke(["streamlit", "deploy"])
 
     root_path = f"@MockDatabase.MockSchema.streamlit_stage/{STREAMLIT_NAME}"
     assert result.exit_code == 0, result.output
     assert ctx.get_queries() == [
         "create stage if not exists IDENTIFIER('MockDatabase.MockSchema.streamlit_stage')",
-        _put_query("streamlit_app.py", root_path),
-        _put_query("environment.yml", root_path),
-        _put_query("pages/*", f"{root_path}/pages"),
+        _put_query(tmp_dir, "streamlit_app.py", root_path),
+        _put_query(tmp_dir, "environment.yml", root_path),
+        _put_query(tmp_dir, "pages/my_page.py", f"{root_path}/pages"),
         dedent(
             f"""
                 CREATE STREAMLIT IDENTIFIER('MockDatabase.MockSchema.{STREAMLIT_NAME}')
@@ -609,16 +601,18 @@ def test_deploy_all_streamlit_files_not_defaults(
     )
     mock_connector.return_value = ctx
 
-    with project_directory(project_name):
+    with project_directory(project_name) as tmp_dir:
         result = runner.invoke(["streamlit", "deploy"])
 
     root_path = f"@MockDatabase.MockSchema.streamlit_stage/{STREAMLIT_NAME}"
     assert result.exit_code == 0, result.output
     assert ctx.get_queries() == [
         "create stage if not exists IDENTIFIER('MockDatabase.MockSchema.streamlit_stage')",
-        _put_query("main.py", root_path),
-        _put_query("streamlit_environment.yml", root_path),
-        _put_query("streamlit_pages/*", f"{root_path}/streamlit_pages"),
+        _put_query(tmp_dir, "main.py", root_path),
+        _put_query(tmp_dir, "streamlit_environment.yml", root_path),
+        _put_query(
+            tmp_dir, "streamlit_pages/first_page.py", f"{root_path}/streamlit_pages"
+        ),
         dedent(
             f"""
             CREATE STREAMLIT IDENTIFIER('MockDatabase.MockSchema.{STREAMLIT_NAME}')
@@ -675,10 +669,10 @@ def test_deploy_streamlit_main_and_pages_files_experimental(
             return_value=enable_streamlit_no_checkouts,
         ),
     ):
-        with project_directory(project_name) as pdir:
+        with project_directory(project_name) as tmp_dir:
             if project_name == "example_streamlit_v2":
                 alter_snowflake_yml(
-                    pdir / "snowflake.yml",
+                    tmp_dir / "snowflake.yml",
                     parameter_path="entities.test_streamlit.artifacts",
                     value=["streamlit_app.py", "environment.yml", "pages"],
                 )
@@ -711,9 +705,9 @@ def test_deploy_streamlit_main_and_pages_files_experimental(
                """
             ).strip(),
             post_create_command,
-            _put_query("streamlit_app.py", root_path),
-            _put_query("environment.yml", f"{root_path}"),
-            _put_query("pages/*", f"{root_path}/pages"),
+            _put_query(tmp_dir, "streamlit_app.py", root_path),
+            _put_query(tmp_dir, "environment.yml", f"{root_path}"),
+            _put_query(tmp_dir, "pages/my_page.py", f"{root_path}/pages"),
             "select system$get_snowsight_host()",
             "select current_account_name()",
         ]
@@ -771,19 +765,17 @@ def test_deploy_streamlit_main_and_pages_files_experimental_double_deploy(
     )
     ctx.queries = []
 
-    with project_directory(project_name) as pdir:
+    with project_directory(project_name) as tmp_dir:
         if project_name == "example_streamlit_v2":
             alter_snowflake_yml(
-                pdir / "snowflake.yml",
+                tmp_dir / "snowflake.yml",
                 parameter_path="entities.test_streamlit.artifacts",
                 value=["streamlit_app.py", "environment.yml", "pages"],
             )
         result2 = runner.invoke(["streamlit", "deploy", "--experimental"])
 
-    assert result2.exit_code == 0, result2.output
-
     root_path = f"@streamlit/MockDatabase.MockSchema.{STREAMLIT_NAME}/default_checkout"
-
+    assert result2.exit_code == 0, result2.output
     # Same as normal, except no ALTER query
     assert ctx.get_queries() == [
         dedent(
@@ -794,9 +786,9 @@ def test_deploy_streamlit_main_and_pages_files_experimental_double_deploy(
             TITLE = 'My Fancy Streamlit'
             """
         ).strip(),
-        _put_query("streamlit_app.py", root_path),
-        _put_query("environment.yml", f"{root_path}"),
-        _put_query("pages/*", f"{root_path}/pages"),
+        _put_query(tmp_dir, "streamlit_app.py", root_path),
+        _put_query(tmp_dir, "environment.yml", f"{root_path}"),
+        _put_query(tmp_dir, "pages/my_page.py", f"{root_path}/pages"),
         "select system$get_snowsight_host()",
         "select current_account_name()",
     ]
@@ -838,7 +830,7 @@ def test_deploy_streamlit_main_and_pages_files_experimental_no_stage(
         "snowflake.cli.api.feature_flags.FeatureFlag.ENABLE_STREAMLIT_VERSIONED_STAGE.is_enabled",
         return_value=enable_streamlit_versioned_stage,
     ):
-        with project_directory(project_name):
+        with project_directory(project_name) as tmp_dir:
 
             result = runner.invoke(["streamlit", "deploy", "--experimental"])
 
@@ -863,9 +855,9 @@ def test_deploy_streamlit_main_and_pages_files_experimental_no_stage(
             """
         ).strip(),
         post_create_command,
-        _put_query("streamlit_app.py", root_path),
-        _put_query("environment.yml", f"{root_path}"),
-        _put_query("pages/*", f"{root_path}/pages"),
+        _put_query(tmp_dir, "streamlit_app.py", root_path),
+        _put_query(tmp_dir, "environment.yml", f"{root_path}"),
+        _put_query(tmp_dir, "pages/my_page.py", f"{root_path}/pages"),
         f"select system$get_snowsight_host()",
         f"select current_account_name()",
     ]
@@ -899,10 +891,10 @@ def test_deploy_streamlit_main_and_pages_files_experimental_replace(
     )
     mock_connector.return_value = ctx
 
-    with project_directory(project_name) as pdir:
+    with project_directory(project_name) as tmp_dir:
         if project_name == "example_streamlit_v2":
             alter_snowflake_yml(
-                pdir / "snowflake.yml",
+                tmp_dir / "snowflake.yml",
                 parameter_path="entities.test_streamlit.artifacts",
                 value=["streamlit_app.py", "environment.yml", "pages/"],
             )
@@ -920,9 +912,9 @@ def test_deploy_streamlit_main_and_pages_files_experimental_replace(
             """
         ).strip(),
         f"ALTER streamlit MockDatabase.MockSchema.{STREAMLIT_NAME} CHECKOUT",
-        _put_query("streamlit_app.py", root_path),
-        _put_query("environment.yml", f"{root_path}"),
-        _put_query("pages/*", f"{root_path}/pages"),
+        _put_query(tmp_dir, "streamlit_app.py", root_path),
+        _put_query(tmp_dir, "environment.yml", f"{root_path}"),
+        _put_query(tmp_dir, "pages/my_page.py", f"{root_path}/pages"),
         f"select system$get_snowsight_host()",
         f"select current_account_name()",
     ]
@@ -1084,7 +1076,7 @@ def test_deploy_streamlit_with_comment_v2(
     )
     mock_connector.return_value = ctx
 
-    with project_directory("example_streamlit_with_comment_v2"):
+    with project_directory("example_streamlit_with_comment_v2") as tmp_dir:
         result = runner.invoke(["streamlit", "deploy", "--replace"])
 
     root_path = f"@MockDatabase.MockSchema.streamlit/test_streamlit_deploy_snowcli"
@@ -1092,9 +1084,9 @@ def test_deploy_streamlit_with_comment_v2(
     assert ctx.get_queries() == [
         f"describe streamlit IDENTIFIER('MockDatabase.MockSchema.test_streamlit_deploy_snowcli')",
         "create stage if not exists IDENTIFIER('MockDatabase.MockSchema.streamlit')",
-        _put_query("streamlit_app.py", root_path),
-        _put_query("pages/*", f"{root_path}/pages"),
-        _put_query("environment.yml", root_path),
+        _put_query(tmp_dir, "streamlit_app.py", root_path),
+        _put_query(tmp_dir, "pages/my_page.py", f"{root_path}/pages"),
+        _put_query(tmp_dir, "environment.yml", root_path),
         dedent(
             f"""
             CREATE OR REPLACE STREAMLIT IDENTIFIER('MockDatabase.MockSchema.test_streamlit_deploy_snowcli')