+++ tests_integration/conftest.py
@@ -24,6 +24,7 @@ from dataclasses import dataclass
 from json import JSONDecodeError
 from pathlib import Path
 from typing import Any, Dict, List, Optional
+from unittest import mock
 from uuid import uuid4
 
 import pytest
@@ -258,3 +259,12 @@ def resource_suffix(request):
     # To generate a suffix that isn't too long or complex, we use originalname, which is the
     # "bare" test function name, without filename, class name, or parameterization variables
     return f"_{uuid4().hex}_{request.node.originalname}"
+
+
+@pytest.fixture
+def enable_snowpark_glob_support_feature_flag():
+    with mock.patch(
+        f"snowflake.cli.api.feature_flags.FeatureFlag.ENABLE_SNOWPARK_GLOB_SUPPORT.is_enabled",
+        return_value=True,
+    ):
+        yield